# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "The Music Scene \u2014 Label Intelligence Platform (Demo)",
    "version": "0.1.0-step7",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /v3/auth/login. Carries a sessionVersion claim; password reset/change revokes previously issued tokens."
      }
    },
...
  },
  "paths": {
...
    "/v3/ai/analyze": {
      "post": {
        "summary": "AI analyze",
        "description": "Structured analysis (growth/roi/tour prompts). Provider-unavailable without key. User-initiated only. Successful responses carry a disclaimer; financial deep analysis requires POST /v3/ai/financial-analysis with explicit opt-in.",
        "tags": [
          "ai"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request / validation failure"
          },
          "401": {
            "description": "Missing, invalid, or revoked bearer token"
          },
          "403": {
            "description": "Authenticated but not permitted (e.g. admin-only, or artist outside access grant)"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v3/ai/providers": {
      "get": {
        "summary": "AI providers",
        "description": "Lists configured AI providers. 503 when none configured.",
        "tags": [
          "ai"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request / validation failure"
          },
          "401": {
            "description": "Missing, invalid, or revoked bearer token"
          },
          "403": {
            "description": "Authenticated but not permitted (e.g. admin-only, or artist outside access grant)"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v3/ai/query": {
      "post": {
        "summary": "AI query",
        "description": "Free-form AI query. Provider-unavailable errors when no key. User-initiated only. Successful responses carry a disclaimer; financial deep analysis requires POST /v3/ai/financial-analysis with explicit opt-in.",
        "tags": [
          "ai"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request / validation failure"
          },
          "401": {
            "description": "Missing, invalid, or revoked bearer token"
          },
          "403": {
            "description": "Authenticated but not permitted (e.g. admin-only, or artist outside access grant)"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v3/exports": {
      "get": {
        "summary": "Data export",
        "description": "Export endpoint. Guarded by authenticateToken + checkExportAccess (pinned by routes.test.js).",
        "tags": [
          "exports"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request / validation failure"
          },
          "401": {
            "description": "Missing, invalid, or revoked bearer token"
          },
          "403": {
            "description": "Authenticated but not permitted (e.g. admin-only, or artist outside access grant)"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v3/reports/generate-all": {
      "post": {
        "summary": "Generate all monthly reports",
        "description": "Admin only. Batch report generation for auto-printing. Requires the admin role (403 otherwise).",
        "tags": [
          "reports"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request / validation failure"
          },
          "401": {
            "description": "Missing, invalid, or revoked bearer token"
          },
          "403": {
            "description": "Authenticated but not permitted (e.g. admin-only, or artist outside access grant)"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v3/reports/monthly/{artistId}/{month}": {
      "get": {
        "summary": "Monthly report PDF",
        "description": "Generates the monthly artist-report PDF. PDF embeds a generation timestamp (size varies).",
        "tags": [
          "reports"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request / validation failure"
          },
          "401": {
            "description": "Missing, invalid, or revoked bearer token"
          },
          "403": {
            "description": "Authenticated but not permitted (e.g. admin-only, or artist outside access grant)"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/v3/ai/financial-analysis": {
      "post": {
        "summary": "AI financial analysis (explicit opt-in)",
        "description": "The ONLY finance-dedicated AI endpoint. Requires acknowledgeNotAdvice: true in the body - an explicit per-request opt-in that this is not financial advice. Successful responses carry a disclaimer and user-responsibility statement. Financial-data posture (2026-09-28, FINANCIAL_DATA_POLICY.md): the platform is a lens on the label's own sales/statements, not a financial custodian; payment integrations are read-only and opt-in (Stripe Connect, TEST MODE ONLY - platform subscription billing and label direct sales are separate). Money is integer cents with ISO currency codes. AI financial analysis is strictly user-initiated and opt-in: scheduled monthly reports never invoke AI by default, and dedicated AI financial analysis requires an explicit per-request acknowledgment that it is not financial advice.",
        "tags": [
          "AI"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Prompt required, or opt-in acknowledgment missing"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "AI is unavailable"
          }
        }
      }
    },
...
  },
...
}